Improvements to allow full testing w/psycopg3.
Also fixes some instances where parameterization of values is not
supported. A surprising one was IS/IS NOT NULL where "NULL" is not
allowed to be parameterized (None converts to NULL with psycopg2).
We'll just explicitly specify the sql literal NULL for these
comparisons. This change affects the other implementations as well.